Updates store and user endpoints
Refactors store endpoints for clarity and consistency.

Modifies user endpoints to use PathVariable for store ID.

Changes DtoC5 to use Float for average daily sales.

Removes password hashing for simplicity.

+++ lab1/src/main/java/com/example/demo/Dtos/DtoC5.java
@@ -3,10 +3,10 @@ package com.example.demo.Dtos;
 public class DtoC5 {
     private String month_name;
     private Integer year;
-    private Integer average_daily_sales;
+    private Float average_daily_sales;
     private Integer difference_from_previous_month;
 
-    public DtoC5(String month_name, Integer year, Integer average_daily_sales, Integer difference_from_previous_month) {
+    public DtoC5(String month_name, Integer year, Float average_daily_sales, Integer difference_from_previous_month) {
         this.month_name = month_name;
         this.year = year;
         this.average_daily_sales = average_daily_sales;
@@ -19,7 +19,7 @@ public class DtoC5 {
     public Integer getYear() {
         return year;
     }
-    public Integer getAverage_daily_sales() {
+    public Float getAverage_daily_sales() {
         return average_daily_sales;
     }
     public Integer getDifference_from_previous_month() {
@@ -32,7 +32,7 @@ public class DtoC5 {
     public void setYear(Integer year) {
         this.year = year;
     }
-    public void setAverage_daily_sales(Integer average_daily_sales) {
+    public void setAverage_daily_sales(Float average_daily_sales) {
         this.average_daily_sales = average_daily_sales;
     }
     public void setDifference_from_previous_month(Integer difference_from_previous_month) {
